requirements.txt
# PDF Processing
pdfplumber>=0.7.0
PyPDF2>=3.0.0

# NLP and Summarization (simplified)
nltk>=3.8.0

# Slide Creation
python-pptx>=0.6.21
pillow>=10.0.0

# Video Processing
moviepy>=1.0.0
imageio-ffmpeg>=0.4.0

# Text-to-Speech
pyttsx3>=2.90

# Utilities
pyyaml>=6.0.0
click>=8.0.0

# Development and Testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
